requirements.txt
bleach==6.3.0
bokeh==3.8.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorcet==3.1.0
contourpy==1.3.3
cycler==0.12.1
dask==2026.1.2
datashader==0.18.2
fonttools==4.61.1
fsspec==2026.2.0
holoviews==1.22.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
linkify-it-py==2.1.0
llvmlite==0.46.0
locket==1.0.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
multipledispatch==1.0.0
narwhals==2.17.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
panel==1.8.9
param==2.3.2
partd==1.4.2
pillow==12.1.1
pyct==0.6.0
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyviz_comms==3.0.6
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tifffile==2026.3.3
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
typing_extensions==4.15.0
uc-micro-py==2.0.0
umap-learn==0.5.11
urllib3==2.6.3
webencodings==0.5.1
xarray==2026.2.0
xyzservices==2025.11.0
